Updated to support the new data format Renamed to dockerservice
